+++ src/multi_model.rs
@@ -71,16 +71,30 @@ pub struct MultiModelViewerGaussianBuffers<G: GaussianPod = DefaultGaussianPod> 
     pub radix_sort_indirect_args_buffer: RadixSortIndirectArgsBuffer,
     pub indirect_indices_buffer: IndirectIndicesBuffer,
     pub gaussians_depth_buffer: GaussiansDepthBuffer,
+    #[cfg(feature = "viewer-selection")]
+    pub selection_buffer: SelectionBuffer,
+    #[cfg(feature = "viewer-selection")]
+    pub invert_selection_buffer: selection::PreprocessorInvertSelectionBuffer,
 }
 
 impl<G: GaussianPod> MultiModelViewerGaussianBuffers<G> {
     /// Create a new viewer Gaussian buffers.
     pub fn new(device: &wgpu::Device, gaussians: &Gaussians) -> Self {
+        Self::new_with(device, GaussiansBuffer::<G>::DEFAULT_USAGES, gaussians)
+    }
+
+    /// Create a new viewer Gaussian buffers with all the extra options.
+    pub fn new_with(
+        device: &wgpu::Device,
+        gaussians_buffer_usage: wgpu::BufferUsages,
+        gaussians: &Gaussians,
+    ) -> Self {
         log::debug!("Creating model transform buffer");
         let model_transform_buffer = ModelTransformBuffer::new(device);
 
         log::debug!("Creating gaussians buffer");
-        let gaussians_buffer = GaussiansBuffer::new(device, &gaussians.gaussians);
+        let gaussians_buffer =
+            GaussiansBuffer::new_with_usage(device, &gaussians.gaussians, gaussians_buffer_usage);
 
         log::debug!("Creating indirect args buffer");
         let indirect_args_buffer = IndirectArgsBuffer::new(device);
@@ -96,6 +110,18 @@ impl<G: GaussianPod> MultiModelViewerGaussianBuffers<G> {
         let gaussians_depth_buffer =
             GaussiansDepthBuffer::new(device, gaussians.gaussians.len() as u32);
 
+        #[cfg(feature = "viewer-selection")]
+        let selection_buffer = {
+            log::debug!("Creating selection buffer");
+            SelectionBuffer::new(device, gaussians.gaussians.len() as u32)
+        };
+
+        #[cfg(feature = "viewer-selection")]
+        let invert_selection_buffer = {
+            log::debug!("Creating invert selection buffer");
+            selection::PreprocessorInvertSelectionBuffer::new(device)
+        };
+
         Self {
             model_transform_buffer,
             gaussians_buffer,
@@ -103,6 +129,10 @@ impl<G: GaussianPod> MultiModelViewerGaussianBuffers<G> {
             radix_sort_indirect_args_buffer,
             indirect_indices_buffer,
             gaussians_depth_buffer,
+            #[cfg(feature = "viewer-selection")]
+            selection_buffer,
+            #[cfg(feature = "viewer-selection")]
+            invert_selection_buffer,
         }
     }
 
@@ -126,6 +156,18 @@ impl<G: GaussianPod> MultiModelViewerGaussianBuffers<G> {
         log::debug!("Creating gaussians depth buffer");
         let gaussians_depth_buffer = GaussiansDepthBuffer::new(device, count as u32);
 
+        #[cfg(feature = "viewer-selection")]
+        let selection_buffer = {
+            log::debug!("Creating selection buffer");
+            SelectionBuffer::new(device, count as u32)
+        };
+
+        #[cfg(feature = "viewer-selection")]
+        let invert_selection_buffer = {
+            log::debug!("Creating invert selection buffer");
+            selection::PreprocessorInvertSelectionBuffer::new(device)
+        };
+
         Self {
             model_transform_buffer,
             gaussians_buffer,
@@ -133,6 +175,10 @@ impl<G: GaussianPod> MultiModelViewerGaussianBuffers<G> {
             radix_sort_indirect_args_buffer,
             indirect_indices_buffer,
             gaussians_depth_buffer,
+            #[cfg(feature = "viewer-selection")]
+            selection_buffer,
+            #[cfg(feature = "viewer-selection")]
+            invert_selection_buffer,
         }
     }
 
@@ -184,6 +230,10 @@ impl MultiModelViewerBindGroups {
             &gaussian_buffers.radix_sort_indirect_args_buffer,
             &gaussian_buffers.indirect_indices_buffer,
             &gaussian_buffers.gaussians_depth_buffer,
+            #[cfg(feature = "viewer-selection")]
+            &gaussian_buffers.selection_buffer,
+            #[cfg(feature = "viewer-selection")]
+            &gaussian_buffers.invert_selection_buffer,
         );
         let radix_sorter = radix_sorter.create_bind_groups(
             device,
@@ -236,7 +286,10 @@ impl<G: GaussianPod, K: Hash + std::cmp::Eq> MultiModelViewer<G, K> {
         Self::new_with(device, texture_format, None)
     }
 
-    /// Create a new viewer with all options.
+    /// Create a new viewer with all extra options.
+    ///
+    /// More specifically, you can specify:
+    /// - `depth_stencil`: The optional depth stencil state for the renderer.
     pub fn new_with(
         device: &wgpu::Device,
         texture_format: wgpu::TextureFormat,
@@ -268,8 +321,28 @@ impl<G: GaussianPod, K: Hash + std::cmp::Eq> MultiModelViewer<G, K> {
     }
 
     /// Insert a new model to the viewer.
-    pub fn insert_model(&mut self, device: &wgpu::Device, key: K, gaussians: &Gaussians) {
-        let gaussian_buffers = MultiModelViewerGaussianBuffers::new(device, gaussians);
+    pub fn insert_model(
+        &mut self,
+        device: &wgpu::Device,
+        key: K,
+        gaussians: &Gaussians,
+    ) -> Option<MultiModelViewerModel<G>> {
+        self.insert_model_with(device, key, GaussiansBuffer::<G>::DEFAULT_USAGES, gaussians)
+    }
+
+    /// Insert a new model to the viewer with all extra options.
+    ///
+    /// More specifically, you can specify:
+    /// - `gaussians_buffer_usage`: The usage for the gaussians buffer.
+    pub fn insert_model_with(
+        &mut self,
+        device: &wgpu::Device,
+        key: K,
+        gaussians_buffer_usage: wgpu::BufferUsages,
+        gaussians: &Gaussians,
+    ) -> Option<MultiModelViewerModel<G>> {
+        let gaussian_buffers =
+            MultiModelViewerGaussianBuffers::new_with(device, gaussians_buffer_usage, gaussians);
         let bind_groups = MultiModelViewerBindGroups::new(
             device,
             &self.preprocessor,
@@ -284,12 +357,12 @@ impl<G: GaussianPod, K: Hash + std::cmp::Eq> MultiModelViewer<G, K> {
                 gaussian_buffers,
                 bind_groups,
             },
-        );
+        )
     }
 
     /// Remove a model from the viewer.
-    pub fn remove_model(&mut self, key: &K) {
-        self.models.remove(key);
+    pub fn remove_model(&mut self, key: &K) -> Option<MultiModelViewerModel<G>> {
+        self.models.remove(key)
     }
 
     /// Update the camera.
@@ -316,12 +389,13 @@ impl<G: GaussianPod, K: Hash + std::cmp::Eq> MultiModelViewer<G, K> {
         pos: Vec3,
         rot: Quat,
         scale: Vec3,
-    ) {
+    ) -> Result<(), MultiModelViewerAccessError> {
         self.models
             .get_mut(key)
-            .expect("model not found")
+            .ok_or(MultiModelViewerAccessError::ModelNotFound)?
             .gaussian_buffers
             .update_model_transform(queue, pos, rot, scale);
+        Ok(())
     }
 
     /// Update the model transform with [`ModelTransformPod`].
@@ -330,12 +404,13 @@ impl<G: GaussianPod, K: Hash + std::cmp::Eq> MultiModelViewer<G, K> {
         queue: &wgpu::Queue,
         key: &K,
         pod: &ModelTransformPod,
-    ) {
+    ) -> Result<(), MultiModelViewerAccessError> {
         self.models
             .get_mut(key)
-            .expect("model not found")
+            .ok_or(MultiModelViewerAccessError::ModelNotFound)?
             .gaussian_buffers
             .update_model_transform_with_pod(queue, pod);
+        Ok(())
     }
 
     /// Update the Gaussian transform.
@@ -373,15 +448,17 @@ impl<G: GaussianPod, K: Hash + std::cmp::Eq> MultiModelViewer<G, K> {
         encoder: &mut wgpu::CommandEncoder,
         texture_view: &wgpu::TextureView,
         keys: &[&K],
-    ) -> Result<(), MultiModelViewerRenderError> {
-        if keys.len() != self.models.len() {
-            return Err(MultiModelViewerRenderError::ModelKeyCountMismatch {
-                model_count: self.models.len(),
-                key_count: keys.len(),
-            });
-        }
+    ) -> Result<(), MultiModelViewerAccessError> {
+        let models = keys
+            .iter()
+            .map(|key| {
+                self.models
+                    .get(key)
+                    .ok_or(MultiModelViewerAccessError::ModelNotFound)
+            })
+            .collect::<Result<Vec<_>, _>>()?;
 
-        for model in self.models.values() {
+        for model in models.iter() {
             self.preprocessor.preprocess(
                 encoder,
                 &model.bind_groups.preprocessor,
@@ -412,8 +489,7 @@ impl<G: GaussianPod, K: Hash + std::cmp::Eq> MultiModelViewer<G, K> {
                 timestamp_writes: None,
             });
 
-            for key in keys.iter() {
-                let model = &self.models.get(key).expect("model not found");
+            for model in models.iter() {
                 self.renderer.render_with_pass(
                     &mut render_pass,
                     &model.bind_groups.renderer,